requirements.txt
fastapi
uvicorn
requests
mcp
pyyaml
python-hcl2
pytest
